Moved all relevant code to model_endpoints.py
