--- mlrun/monitoring/constants.py
@@ -1,10 +1,0 @@
-DEFAULT_CONTAINER = "monitoring"
-ENDPOINTS_TABLE = "endpoints"
-
-ENDPOINT_EVENT_LOG_TABLE = "endpoint_event_log"
-ENDPOINT_STATE_LOG_TABLE = "endpoint_state_log"
-ENDPOINT_ALERT_LOG_TABLE = "endpoint_alert_log"
-ENDPOINT_ALERT_LOG_STREAM = "endpoint_alert_log_stream"
-
-ISO_8601 = "%Y-%m-%d %H:%M:%S.%f"
-ISO_8601_NO_MILLIES = "%Y-%m-%d %H:%M:%S"
